+++ mediation/admob/interstitial/MTRGAdmobRewardedVideoCustomEvent.h
@@ -1,0 +1,13 @@
+//
+//  MTRGAdmobRewardedVideoCustomEvent.h
+//  myTargetSDKAdmobMediation
+//
+//  Created by Andrey Seredkin on 06.10.16.
+//  Copyright © 2016 Mail.ru Group. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface MTRGAdmobRewardedVideoCustomEvent : NSObject
+
+@end