+++ mediation/mopub/interstitial/MTRGMopubRewardedVideoCustomEvent.h
@@ -1,0 +1,13 @@
+//
+//  MTRGMopubRewardedVideoCustomEvent.h
+//  myTargetSDKMopubMediation
+//
+//  Created by Andrey Seredkin on 05.10.16.
+//  Copyright (c) 2016 Mail.ru Group. All rights reserved.
+//
+
+#import "MPRewardedVideoCustomEvent.h"
+
+@interface MTRGMopubRewardedVideoCustomEvent : MPRewardedVideoCustomEvent
+
+@end